feat: style top bar section
I found this to be brilliant.

The top bar section has content on the right and on the dead center, but
not on the right, which is pretty hard to do. Luckily, we can do this
with grid.

What we're doing is to set three columns. The second column, which is
the anchor list, will occupy the least amount of space, and the first
and last columns will occupy THE REST OF THE SPACE. And this is key,
since it forces the second column to be AT THE CENTER

Similar to mx-auto, but with grid. I liked it.