style: set max width to header content
We did this because according to the Figma design, this is smaller than
most of the components on the page. When you're setting max widths, it's
a good idea to set it in terms of pixels. It'll work fine on very large
screens, and it will prevent your content from stretching through the
whole screen width