style: add extra spacing on top of header title
In this case, we used vh units because the surrounding elements are
spaced with vh units, and this will cause to space elements WITH RESPECT
to the user's screen size. This means that using vh units will guarantee
that elements on the header section will take that into account and
align well.

If you were to leave a fixed space, users with large screens would
suffer a lot