style: add styles to tagline dash
In our solution, we made use of multiple dashes on the content.

On Kyle's solution, it just made the before element a block element so
that he can CHANGE THE SIZE OF IT (this is key; you can use either block
or inline-block elements if you desire to change its size)

Now, by doing that, he changed its width and height, thus being able to
change its background color

I liked that solution :)

Yup, you could've specified the ::before element as an inline-block
instead of a block element. There are different solutions :)