style: change title's font size at a certain width
The reason for this is that at around 1200px width, the header's title
gets pretty close to both the left and right hand side of the screen
without getting to the point where's actually touching it. If we were to
shrink down the screen's width a little bit more, the title would end up
wrapping to three lines, which looks visually bad.

In order to avoid that, we reduced its font size a little bit, from
5.5rem to 4rem, which, in the end, looks pretty good

It doesn't touch the right or left hand side of the screen