style: change header's title font size
We corrected it for viewport widths less than 1200px, but it still looks
huge when we approach widths of around less than 800px, so we reduced
the font size from 4rem to 3.25 rem when it's about 800px

Yup, that gives a really good feeling of the website. I do like it.