feat: reduce margin at the bottom
If we add content below the fade container, it'll start right after the
fade effect, but we don't want that; we want the content to be somewhat
on the fade.

In order to correct that, we'll add a negative margin at the bottom.
It'll help placing the next content over the fade effect, which is
exactly what we want, according to the Figma design.