feat: change fade background's height
If you set the height of the fade container to be below 50vh (47vh, for
instance), you can see the header's background image is bigger than the
fade container by 3vh

We can correct it by setting the container's height to be 50vh. Great.
The thing is that SOMETIMES you could see a slight line that corresponds
to the background image. This means that even we're covering the
overflowing background image piece, it won't completely cover it
(depends on your screen).

In order to correct that, we make the container's height SLIGHTLY (I
mean slightly) bigger, just by 1px so that we don't see that annoying
line at the bottom

I loved that detail.